+++ UnityUtility/Assets/Public/Editor/CustomAttributes/AttributeDrawers/ButtonAttributeDrawer.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
 using UnityEditor;
+using UnityEditor.UIElements;
 using UnityEngine;
 using UnityEngine.UIElements;
 
@@ -11,26 +12,14 @@ namespace UnityUtility.CustomAttributes.Editor
     [CustomPropertyDrawer(typeof(ButtonAttribute))]
     public class ButtonAttributeDrawer : PropertyDrawer
     {
+        private ButtonAttribute m_target = null;
+
         private bool m_isInit = false;
 
         private MethodInfo m_targetMethod = null;
         private object m_targetObject = null;
-        private ButtonAttribute m_target = null;
 
-        public override VisualElement CreatePropertyGUI(SerializedProperty property)
-        {
-            VisualElement container = new VisualElement();
-
-            InitIfNeeded(property);
-            
-            Button b = new Button(() => InvokeMethod());
-            b.text = m_target.DisplayName;
-            container.Add(b);
-            container.Add(base.CreatePropertyGUI(property));
-
-            return container;
-        }
-
+        #region IMGUI
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
             InitIfNeeded(property);
@@ -48,6 +37,24 @@ namespace UnityUtility.CustomAttributes.Editor
         {
             return EditorGUIUtility.singleLineHeight + base.GetPropertyHeight(property, label);
         }
+        #endregion
+
+        #region UIElements
+        public override VisualElement CreatePropertyGUI(SerializedProperty property)
+        {
+            VisualElement container = new VisualElement();
+
+            InitIfNeeded(property);
+            
+            Button b = new Button(() => InvokeMethod());
+            b.text = m_target.DisplayName;
+            container.Add(b);
+            container.Add(new PropertyField(property));
+
+            return container;
+        }
+        #endregion
+
 
         private void InitIfNeeded(SerializedProperty property)
         {