+++ UnityUtility/Assets/Public/Editor/CustomAttributes/AttributeDrawers/LabelAttributeDrawer.cs
@@ -1,53 +1,90 @@
-using UnityEngine;
 using UnityEditor;
-using System;
 using UnityEngine.UIElements;
 using UnityEditor.UIElements;
+using UnityUtility.Utils.Editor;
 
 namespace UnityUtility.CustomAttributes.Editor
 {
     [CustomPropertyDrawer(typeof(LabelAttribute))]
     public class LabelAttributeDrawer : PropertyDrawer
     {
-        private VisualElement m_drawnProperty;
+        private PropertyField m_drawnProperty;
+        private bool m_partOfArray;
 
-        #region VisualElement
+        #region UIElements
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
+            m_partOfArray = property.IsPropertyPartOfArray(out SerializedProperty arrayProperty, out int index);
             m_drawnProperty = new PropertyField(property);
+            m_drawnProperty.name = "DrawnProperty";
+
             m_drawnProperty.RegisterCallback<AttachToPanelEvent>(OnAttachToPanel);
             return m_drawnProperty;
         }
 
         private void OnAttachToPanel(AttachToPanelEvent evt)
         {
+            m_drawnProperty.UnregisterCallback<AttachToPanelEvent>(OnAttachToPanel);
             if (attribute is LabelAttribute labelAttribute)
             {
-                Label propertyLabel = m_drawnProperty.Q<Label>();
+                Label propertyLabel = null;
+                if (m_partOfArray)
+                {
+                    PropertyField drawnArray = FindInParent<PropertyField>(m_drawnProperty.parent.parent);
+                    if (drawnArray == null)
+                    {
+                        return;
+                    }
+                    propertyLabel = drawnArray.Q<Label>();
+                }
+                else
+                {
+                    propertyLabel = m_drawnProperty.Q<Label>();
+                }
+
+                if (propertyLabel == null)
+                {
+                    return;
+                }
+
+                // Label name
                 if (!string.IsNullOrEmpty(labelAttribute.OverrideName))
                 {
                     propertyLabel.text = labelAttribute.OverrideName;
                 }
 
-                FontStyle labelStyle = FontStyle.Normal;
-                if (labelAttribute.Bold && labelAttribute.Italic)
+                // Label font style
+                propertyLabel.style.unityFontStyleAndWeight = AttributeUtils.GetFontStyle(labelAttribute.Bold, labelAttribute.Italic);
+
+                // Label font size
+                propertyLabel.style.fontSize = labelAttribute.FontSize;
+            }
+        }
+
+
+        private T FindInParent<T>(VisualElement v, string name = null) where T : VisualElement
+        {
+            if (v == null)
+            {
+                return null;
+            }
+
+            if (v is T matchingParent)
+            {
+                if (!string.IsNullOrEmpty(name))
                 {
-                    labelStyle = FontStyle.BoldAndItalic;
+                    if (matchingParent.name == name)
+                    {
+                        return matchingParent;
+                    }
                 }
                 else
                 {
-                    if (labelAttribute.Bold)
-                    {
-                        labelStyle = FontStyle.Bold;
-                    }
-                    else if (labelAttribute.Italic)
-                    {
-                        labelStyle = FontStyle.Italic;
-                    }
+                    return matchingParent;
                 }
-                propertyLabel.style.unityFontStyleAndWeight = labelStyle;
-
             }
+
+            return FindInParent<T>(v.parent, name);
         }
         #endregion
     }