+++ UnityUtility/Assets/Public/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.4] - 2024-03-11
+
+### Added
+- `Disable` and `DisableIf` attributes
+- UIElements support for `MinMaxSlider` attribute
+- Font size option for the `Title` and `Label` attributes
+- `Vectors.SqrDistance`
+
 ## [1.8.3] - 2024-03-10
 
 ### Added